Moving the logic of eval_equation_system in numerical.rs
